+++ lib/Dashboard/Widget.php
@@ -84,5 +84,6 @@ class Widget implements IWidget
         Util::addScript('analytics', '3rdParty/chartjs-adapter-moment');
         Util::addScript('analytics', '3rdParty/moment');
         Util::addScript('analytics', '3rdParty/cloner');
+        Util::addScript('analytics', '3rdParty/chartjs-plugin-datalabels.min');
     }
 }