+++ templates/main.php
@@ -25,6 +25,7 @@ Util::addScript('analytics', 'jquery.dataTables.min');
 //Util::addScript('analytics', '3rdParty/Chart.bundle.min');
 Util::addScript('analytics', '3rdParty/chart.min');
 Util::addScript('analytics', '3rdParty/chartjs-adapter-moment');
+Util::addScript('analytics', '3rdParty/chartjs-plugin-datalabels.min');
 Util::addScript('analytics', '3rdParty/moment');
 Util::addScript('analytics', '3rdParty/cloner');
 Util::addScript('analytics', 'dashboard');